feat: Fully implements output clause ready to complet INSERT statement.
